Refactor locations
