--- warehouse-frontend-testing/src/test/java/com/skillstorm/testingComponents/HomePage.java
@@ -1,5 +1,0 @@
-package com.skillstorm.testingComponents;
-
-public class HomePage implements Page {
-    
-}
