--- warehouse-frontend-testing/src/test/java/com/skillstorm/testingComponents/ItemsPage.java
@@ -1,5 +1,0 @@
-package com.skillstorm.testingComponents;
-
-public class ItemsPage implements Page {
-    
-}
