--- warehouse-frontend-testing/src/test/java/com/skillstorm/testingComponents/LoginPage.java
@@ -1,5 +1,0 @@
-package com.skillstorm.testingComponents;
-
-public class LoginPage implements Page {
-    
-}
