--- warehouse-frontend-testing/src/test/java/com/skillstorm/testingComponents/Page.java
@@ -1,7 +1,0 @@
-package com.skillstorm.testingComponents;
-
-import java.util.List;
-
-public interface Page {
-    
-}
