+++ warehouse-frontend-testing/src/test/java/com/skillstorm/testingComponents/pages/Page.java
@@ -1,0 +1,5 @@
+package com.skillstorm.testingComponents.pages;
+
+public interface Page {
+    
+}
