+++ warehouse-frontend-testing/src/test/java/com/skillstorm/testingComponents/pages/concretePages/ItemsPage.java
@@ -1,0 +1,7 @@
+package com.skillstorm.testingComponents.pages.concretePages;
+
+import com.skillstorm.testingComponents.pages.Page;
+
+public class ItemsPage implements Page {
+    
+}
