+++ warehouse-frontend-testing/src/test/java/com/skillstorm/testingComponents/pages/concretePages/SignupPage.java
@@ -1,0 +1,7 @@
+package com.skillstorm.testingComponents.pages.concretePages;
+
+import com.skillstorm.testingComponents.pages.Page;
+
+public class SignupPage implements Page {
+    
+}
